Moving camera logic directly into server.py for upcoming flight testing.
Ideally we separate camera code into its own modules/camera.py folder, but unfortunately this needs to query vehicle state when it sends vehicle information
json back with the image. We could send the current vehicle_state prior to calling on camera code, but this will not work because it is parallelized on a separate
thread and image capture will take time. When we send, it will not contain the most recent state. Depending on our flight speed this could be off by several metres
which is unacceptable for target approximation and mapping.